New get_shark_path utility function
It collapses the two values "path" and "config_file" into a single
"path" value to be able to specify a "shark path" either directly or
indirectly by specifying a configuration file where there "shark path"
can be derived from.

Both arguments cannot be given together. When path is given, it's used
"as is". When the config file is given it is loaded and the shark path
is composed using the appropriate bits from the configuration file. When
none is given '.' is used as path. In all cases the final path returned
by the function is checked for read access.

This function is meant only for internal use, and therefore doesn't need
to be exported, or a manual page written about it. To achieve this I've
changed the list of exported symbols in the NAMESPACE file to explicitly
export individual functions instead of exporting with a very generic
pattern that matches almost every single function in the package.

Reading a configuration file is done using the configr package, so I'm
adding that to the Imports list.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>