Renaming current write.SED function
This change of name is to accommodate for a new write.SED function that
will output the SED using different formats depending on the extension
of the filename given to it.

Nobody seemed to be using write.SED anyway (at least internally), so
this will hopefully not break anyone's code.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>