New write.SED.CSV and write.SED functions
The former is like write.SED.hdf5, but outputs the SED into a CSV file.
This is what is already in genShark and genSting.

The latter is the generic function that users can call, and calls either
write.SED.hdf5 or write.SED.csv based on the extension present in the
filename given by the user.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>