Add option to redirect cluster worker's output
By default a cluster created via makeCluster() redirects all worker's
ouput to /dev/null. This commit adds a new "children_output" option to
the genSting() call to give users the flexibility of where these streams
should go to. By default this new option defaults to /dev/null as well,
therefore matching the default behavior of makeCluster(). It is noted in
makeCluster's help page that in some platforms using "" or "/dev/tty"
will produce the worker's output to be redirected to the master process'
terminal output.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>