Add an extra level of parallelism
Until now the genSting() function works by processing subvolume
snapshots in parallel, and then all galaxies within a snapshot in
serial. Once a snapshot's processing is finished (i.e., all its galaxies
have been processed), the results are reported back, written into the
temporary files, and the next snapshot is processed.

Although in principle this is enough to make use of a machine's
resources, it still has the drawback that the results for a given
snapshot might take a long time to be written to disk. This is a problem
in HPC systems where walltimes might need to be short (e.g., because the
job queue will otherwise not schedule one's jobs), or where jobs might
be unexpectedly cancelled by the system scheduler. In these situations
it would be better to have results written to disk more periodically
instead of collecting many of them in memory before doing a dump.

This commit addressed this shortcoming by introducing a new level of
parallelism in genSting(). Instead of offering parallelism at the
subvolume level by running snapshots in parallel, we now offer
parallelism at the snapshot level by processing galaxies in parallel
within a snapshot. This is done by creating a new cluster within each of
the first cluster's workers, which is then used to distribute the
processing of the snapshot's galaxies. With this new level of
parallelism one can then offer higher parallelism at the snapshot level,
finishing a snapshot's processing work within smaller walltimes, and
therefore dumping a snapshot's results more promptly to disk, ensuring
progress of the overall subvolume's processing needs.

This change has an effect on the signature of genSting(): instead of
having a "cores" option it now has two "cores_per_subvolume" and
"cores_per_snapshot" options, which multiplied together yield the total
number of workers that will process galaxies in parallel.

Another change, slightly unintended, was the change of the getSFHfull
function signature to match that of genSting (i.e., use
cores_per_subvolume rather than simply cores). This function is not used
at the moment, but I'd rather err in the side of caution.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>